requirements.txt
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
exceptiongroup==1.2.2
Flask==3.0.3
greenlet==3.0.3
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mercadopago==2.2.3
mock-alchemy==0.2.6
ngrok==1.4.0
packaging==24.1
pluggy==1.5.0
PyJWT==2.9.0
PyMySQL==1.1.1
pytest==8.3.2
python-dotenv==1.0.1
requests==2.32.3
secure-smtplib==0.1.1
six==1.16.0
SQLAlchemy==2.0.32
tomli==2.0.1
typing_extensions==4.12.2
uni-slugify==0.1.4
Unidecode==1.3.8
urllib3==2.2.3
Werkzeug==3.0.3
